+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Config changes
 
 ### Added
+- Added `mode` config for FastAPI which now supports both `asgi` and `wsgi`.
 - The ability to enable JWT creation with session management, this allows easier integration with services that require JWT based authentication: https://github.com/supertokens/supertokens-core/issues/250
 - You can do BaseRequest.request to get the original request object. Fixes #61